+++ wiki/Changelog.md
@@ -6,6 +6,7 @@ Terminal features
   * Support visible space indication.
   * Fixed insert mode in single-width character mode (#964).
   * New CSI > 0 q to report terminal name and version (#881).
+  * New CSI 2/1 SP Z (ECMA-48 PEC) as character attribute to enforce single/double width (#615, #979, #938, #638, ~#88, ~#671).
 
 Character encoding
   * Enhanced and documented GB18030 support (mintty/wsltty#224).